move packages around a bit
