+++ src/main/kotlin/de/cetvericov/photodump/images/persistence/entity/ImageEntity.kt
@@ -1,4 +1,4 @@
-package de.cetvericov.photodump.persistence.entity
+package de.cetvericov.photodump.images.persistence.entity
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
