+++ src/main/kotlin/de/cetvericov/photodump/images/persistence/repository/ImageRepository.kt
@@ -1,6 +1,6 @@
-package de.cetvericov.photodump.persistence.repository
+package de.cetvericov.photodump.images.persistence.repository
 
-import de.cetvericov.photodump.persistence.entity.ImageEntity
+import de.cetvericov.photodump.images.persistence.entity.ImageEntity
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 
 interface ImageRepository: ReactiveCrudRepository<ImageEntity, Long>
